Improve serial port handling and consolidate classes
