+++ bundles/org.openhab.binding.rme/src/main/java/org/openhab/binding/rme/internal/RMEThingConfiguration.java
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2010-2023 Contributors to the openHAB project
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.openhab.binding.rme.internal;
+
+/**
+ * @author Karel Goderis - Initial contribution
+ */
+public class RMEThingConfiguration {
+
+    public String port;
+    public Integer baudrate;
+    public Integer buffer;
+}
